oracle/requirements.txt
praw==7.1.0
prawcore==1.5.0
pandas==1.1.0
textblob==0.15.3
flashtext==2.7
psycopg2-binary==2.8.6
